Changes related to moving integration tests to TAP from nodeunit.
