+++ test/helper.js
@@ -93,6 +93,7 @@ function createServer(callback) {
                         callback(null, arg);
                 }
         });
+        console.log("done");
 }
 
 function zkMkdirP(dpath, cb) {
